feat(suite-native): add LocationPickers as reusable
