+++ suite-native/trading-residence/src/components/TradingLocationPickers.tsx
@@ -1,0 +1,37 @@
+import { isCountrySubdivisionRequired } from '@suite-common/trading';
+import { type CountryChangeContext } from '@suite-native/analytics';
+import { useFormContext } from '@suite-native/forms';
+
+import { CountryOfResidencePicker } from './CountrySheet/CountryOfResidencePicker';
+import { CountrySubdivisionPicker } from './CountrySheet/CountrySubdivisionPicker';
+import { type TradingLocationFormValues } from '../types/tradingLocationForm';
+
+export type TradingLocationPickersProps = {
+    context: CountryChangeContext;
+    testID: string;
+    hasCountrySubdivisionBottomBorder?: boolean;
+};
+
+export const TradingLocationPickers = ({
+    context,
+    testID,
+    hasCountrySubdivisionBottomBorder = true,
+}: TradingLocationPickersProps) => {
+    const { watch } = useFormContext<TradingLocationFormValues>();
+    const selectedCountry = watch('country');
+    const isSubdivisionRequired = isCountrySubdivisionRequired(selectedCountry?.value);
+
+    return (
+        <>
+            <CountryOfResidencePicker
+                testID={`${testID}/country`}
+                context={context}
+                noBottomBorder={!isSubdivisionRequired}
+            />
+            <CountrySubdivisionPicker
+                testID={`${testID}/country-subdivision`}
+                noBottomBorder={!hasCountrySubdivisionBottomBorder}
+            />
+        </>
+    );
+};
